fix: updateの処理を書く場所が間違っていたので修正
